move checker function into own module
