+++ server/modules/robot/sleep.js
@@ -1,0 +1,4 @@
+const sleep = (milliseconds) => {
+    return new Promise(resolve => setTimeout(resolve, milliseconds))
+}
+module.exports = sleep;
